update build tools to 33.0.1 and update gradle wrapper to 8.2.1. This also updates the volley getter to use the new apis.
