+++ opentelemetry-android-instrumentation/build.gradle.kts
@@ -11,7 +11,7 @@ android {
     namespace = "opentelemetry.rum.instrumentation"
 
     compileSdk = 33
-    buildToolsVersion = "30.0.3"
+    buildToolsVersion = "33.0.1"
 
     defaultConfig {
         minSdk = 21
