+++ splunk-otel-android/build.gradle.kts
@@ -8,7 +8,7 @@ android {
     namespace = "com.splunk.android.rum"
 
     compileSdk = 33
-    buildToolsVersion = "30.0.3"
+    buildToolsVersion = "33.0.1"
 
     defaultConfig {
         minSdk = 21
